refactor: code-reviewer 지적 사항 반영
- LadderCommandService: application → domain/service 패키지로 이동 (네이밍 컨벤션 준수)
- LadderDrawEvent → LadderDrawCommandEvent 이름 변경 (BlockStackingCommandEvent 패턴 통일)
- LadderDrawEventConsumer → LadderDrawCommandEventConsumer 이름 변경
- LadderGameErrorCode: INVALID_POLE_INDEX 추가
- Poles.getPlayer(): IllegalArgumentException → BusinessException(INVALID_POLE_INDEX) 변경
- LadderGame.getRankingsForBroadcast()/getScores(): finalRanks null 방어 추가
