+++ backend/src/main/java/coffeeshout/laddergame/domain/LadderGameErrorCode.java
@@ -10,6 +10,7 @@ import org.springframework.http.HttpStatus;
 public enum LadderGameErrorCode implements ErrorCode {
 
     PLAYER_NOT_FOUND(HttpStatus.NOT_FOUND, "플레이어를 찾을 수 없습니다."),
+    INVALID_POLE_INDEX(HttpStatus.INTERNAL_SERVER_ERROR, "유효하지 않은 기둥 인덱스입니다."),
     ;
 
     private final HttpStatus httpStatus;
